+++ qoqo/src/operations/two_qubit_gate_operations.rs
@@ -513,7 +513,7 @@ pub struct ComplexPMInteraction {
 ///         1 & 0 & 0 & 0 \\\\
 ///         0 & e^{i \phi} & 0 & 0 \\\\
 ///         0 & 0 & e^{i \phi} & 0 \\\\
-///         0 & 0 & 0 & e^{i (2\cdot\phi - \pi)}
+///         0 & 0 & 0 & e^{i (2\cdot\phi + \pi)}
 ///         \end{pmatrix}
 ///
 /// Args:
@@ -536,7 +536,7 @@ pub struct PhaseShiftedControlledZ {
 ///         1 & 0 & 0 & 0 \\\\
 ///         0 & e^{i \phi} & 0 & 0 \\\\
 ///         0 & 0 & e^{i \phi} & 0 \\\\
-///         0 & 0 & 0 & e^{i(2\cdot\phi - \theta)}
+///         0 & 0 & 0 & e^{i(2\cdot\phi + \theta)}
 ///         \end{pmatrix}
 ///
 /// Args: