+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This changelog track changes to the qoqo project starting at version v0.5.0
 
+## 1.1.0-beta.6
+
+* Fixed github pipeline deploy
+
+## 1.1.0-beta.5
+
+* Fixed github pipeline deploy qoqo coverage
+
+## 1.1.0-beta.4
+
+* Fixed compatibility tests
+
+## 1.1.0-beta.3
+
+* Added rich comparison for Python interface of Measurements
+* Added PragmaLoop
+
 ## 1.1.0-beta.1
 
 * Allowed creating PragmaSetStateVector from float or integer numpy arrays.